1.8.6
fix windows build with sin

+++ Navigator/engine/Pipeline.hpp
@@ -604,19 +604,19 @@ void processInput(SDL_Window* /*window*/, int window_width, int window_height) n
     glm::vec3 pos = CAM.position();
 
     if (Options::Camera::EnableHeadBob) {
-        float bob = std::sinf(static_cast<float>(TotalTime::get().seconds()) * Options::Camera::HeadBobFrequency) * Options::Camera::HeadBobIntensity;
+        float bob = std::sin(static_cast<float>(TotalTime::get().seconds()) * Options::Camera::HeadBobFrequency) * Options::Camera::HeadBobIntensity;
         pos.y += bob;
     }
 
     if (Options::Camera::EnableBreathing) {
-        float breath = std::sinf(static_cast<float>(TotalTime::get().seconds()) * 0.8f) * Options::Camera::BreathingIntensity;
+        float breath = std::sin(static_cast<float>(TotalTime::get().seconds()) * 0.8f) * Options::Camera::BreathingIntensity;
         pos.y += breath;
     }
 
     if (Options::Camera::EnableCameraShake) {
         static float trauma = 0.0f;
         trauma *= Options::Camera::ShakeTraumaDecay;
-        float shake = trauma * trauma * (std::sinf(static_cast<float>(TotalTime::get().seconds()) * 10.0f) + std::sinf(static_cast<float>(TotalTime::get().seconds()) * 7.0f));
+        float shake = trauma * trauma * (std::sin(static_cast<float>(TotalTime::get().seconds()) * 10.0f) + std::sin(static_cast<float>(TotalTime::get().seconds()) * 7.0f));
         pos += glm::vec3(shake, shake * 0.5f, 0.0f);
     }
